==> notebook/1.ipynb <==
# Скачиваем информацию о слабах ННР

import requests
import string
import itertools
import concurrent.futures
import time
import os
from tqdm import tqdm
import pandas as pd

### Проверяем все имена слабов формата LLDDDD.

Всего 6760000 вариантов. ~60 часов. Существующие номера сохраняем в файл base.txt

# Создаем директорию для хранения страниц, если она не существует
os.makedirs('D:/Python/coin_pages', exist_ok=True)

def slab_number_generator(start_number):
    letters = string.ascii_uppercase
    digits = string.digits

    start_letter1 = start_number[0].upper()
    start_letter2 = start_number[1].upper()
    start_digit = start_number[2:]

    # Генерируем все возможные комбинации номеров слаба
    for letter1 in letters[letters.index(start_letter1):]:
        for letter2 in letters if letter1 != start_letter1 else letters[letters.index(start_letter2):]:
            for digit_combination in itertools.product(digits, repeat=4):
                slab_number = f"{letter1}{letter2}{''.join(digit_combination)}"
                if slab_number >= start_number.upper():
                    yield slab_number


def check_file_exists(slab_number):
    url = f"https://nreestr.ru/img/all/bpict/{slab_number}r.jpg"
    try:
        response = requests.head(url)
        if response.status_code == 200:
            save_progress('D:/Python/coin_pages/base.txt',f'{slab_number},')
        return response.status_code == 200
    except requests.exceptions.RequestException as e:
        print(f"An error occurred: {e}")
        return False

def save_progress(filename, name):
    with open(filename, 'a') as file:
        file.write(name)


def main(start_number):
    slab_numbers = slab_number_generator(start_number)
    count = 0

    with concurrent.futures.ThreadPoolExecutor(max_workers=10) as executor:
        future_to_slab = {executor.submit(check_file_exists, slab_number): slab_number for slab_number in slab_numbers}

        for future in concurrent.futures.as_completed(future_to_slab):
            slab_number = future_to_slab[future]
            try:
                result = future.result()
                count += 1
                if count % 6000 == 0:
                    print(f"Processed {count} slab numbers")
            except Exception as exc:
                print(f"Slab number {slab_number} generated an exception: {exc}")

if __name__ == "__main__":
    start_number = input("Enter the starting slab number (e.g., vf0000): ").upper()
    main(start_number)

### Скачиваем страницы относящиеся к монетам.

Создаем датафрейм из base.txt . Удаляем дубликаты.

import csv

with open('D:/Python/base.txt', 'r') as f:
    df = pd.DataFrame(f.read().split(','),columns=['number'])
print(df.head())

df.info()

df.pivot_table(index=df['number'].str[:2],values='number',aggfunc='count').describe()

Сохраняем датафрейм в виде csv.

df = df.drop_duplicates()
df.duplicated().sum()
df = df.drop(labels=[121110],axis= 0)
df.to_csv(r'D:\Python\df_base.csv')

df.tail()

Скачиваем с сайта страницы для всех монет в базе. ~3ч.

def download_page(slab_number):
    url = f'https://nreestr.ru/coins/search/{slab_number}';
    try:
        response = requests.get(url)
        if response.status_code == 200:
            # Сохраняем страницу в файл
            filename = f"D:/Python/coin_pages/{slab_number}.html"
            with open(filename, 'w', encoding='utf-8') as file:
                file.write(response.text)
                
            url_r = f"https://nreestr.ru/img/all/bpict/{slab_number}r.jpg"  
            response = requests.get(url_r)   
            reversname = f"D:/Python/coin_pages/{slab_number}r.jpg"
            with open(reversname, 'wb') as file:
                file.write(response.content)    


            url_a = f"https://nreestr.ru/img/all/bpict/{slab_number}a.jpg"  
            response = requests.get(url_a)   
            reversname = f"D:/Python/coin_pages/{slab_number}a.jpg"
            with open(reversname, 'wb') as file:
                file.write(response.content)    

            
            time.sleep(0.2)    
            return f"Saved: {slab_number}"
        else:
            return f"Failed: {slab_number} (Status code: {response.status_code})"
    except Exception as e:
        return f"Error: {slab_number} ({e})"

count = 0
with concurrent.futures.ThreadPoolExecutor(max_workers=10) as executor:
    future_to_slab = {executor.submit(download_page, slab_number): slab_number for slab_number in df['number']}

    for future in concurrent.futures.as_completed(future_to_slab):
        slab_number = future_to_slab[future]
        try:
            result = future.result()
            count += 1
            if count % 200 == 0:
                print(f"Processed {count} slab numbers")
        except Exception as exc:
            print(f"Slab number {slab_number} generated an exception: {exc}")

download_page('YU2656')
