==> notebook/2.ipynb <==
## Извлекаем из html информацию

!pip install beautifulsoup4

from bs4 import BeautifulSoup
import pandas as pd
import time
from tqdm import tqdm
import re
import os

Создаем датафрейм с пустыми колонками.

df = pd.read_csv("D:\Python\df_base.csv",index_col='Unnamed: 0')
df['weight'] = ''
df['date'] = ''
df['year'] = ''
df['condition'] = ''
df['value'] = ''
df['text'] = ''
df.info()

Извлекаем информацию из html.

count = 0
for i in tqdm(df['number']):
    try:
        with open(f"D:\Python\coin_pages\{i}.html", mode='r',encoding='utf-8') as file:
            soup = BeautifulSoup(file, 'html.parser')
            list = soup.find_all('div', 'col-5')
            # Возможность возобновлять процесс.
            if df.loc[df['number']==i,'value'].item() == '':
                count = count+1
                # Обработка исключений, если будет ошибка в индексах списков.
                try:
                    df.loc[df['number']==i,'date'] = list[1].text.strip()
                    df.loc[df['number']==i,'year'] = list[2].text.strip()
                    df.loc[df['number']==i,'condition'] = list[3].text.strip()
                    df.loc[df['number']==i,'weight'] = list[4].text.strip()
                    df.loc[df['number']==i,'text'] = soup.find('div',"mt-3 font-weight-bold").text
                    df.loc[df['number']==i,'value'] = soup.find('div',"col-sm-6 offset-sm-3 col-12").text
                except Exception:
                    print(f'Problem with parsing {i}')
                # Сохраняем прогресс каждую тысячу номеров.    
                if count%1000 == 0:
                    df.to_csv(r'D:\Python\coin_data.csv')
    except Exception:
        print(f'Can not find {i}')
                
df.to_csv(r'D:\Python\coin_data.csv')

df.info()

Сохраняем информацию о монетах с пустыми страницами.

bad_numbers = df.loc[df['weight']=='','number']
bad_numbers.to_csv(r'D:\Python\bad_numbers.csv')

Удаляем из основного датафрейма информацию о монетах с пустыми страницами и сохраняем.

df = df.loc[df['weight']!='']
df['weight'] = df['weight'].astype('float')
df['date'] = pd.to_datetime(df['date'],format="%d.%m.%Y")
df.to_csv(r'D:\Python\coin_data.csv')
df.info()

Удаляем html

html_list = pd.read_csv("D:\Python\df_base.csv",index_col='Unnamed: 0')

for i in tqdm(html_list['number']):
    file_path = f"D:\\Python\\coin_pages\\{i}.html"
    try:
        if os.path.isfile(file_path):
            os.remove(file_path)
    except Exception as e:
        print(f'Ошибка при удалении файла {file_path}. {e}')
