==> notebook/3_final_df.ipynb <==
## Исправляем ошибки и отбираем монеты с оценками по Шелдону.

import pandas as pd
import time
from tqdm import tqdm
import re
import os

df = pd.read_csv("D:\Python\coin_data.csv",index_col='Unnamed: 0')
df.info()

Проверяем все ли загружено.

mising_image = []
for i in tqdm(df['number']):
    if not (os.path.exists(f"D:\Python\coin_pages\{i}a.jpg") and os.path.exists(f"D:\Python\coin_pages\{i}r.jpg")):
        mising_image.append(i)
mising_image

pd.options.display.max_rows = 150
df.pivot_table(index='condition',values='value',aggfunc='count').sort_values(by='value',ascending=False).head(10)

len(df['condition'].unique())

df.loc[df['condition'] == 'XF det.','condition'] = 'ХF det.'
df.loc[df['condition'] == 'UNC  det.','condition'] = 'UNC det.'
df.loc[df['condition'] == 'UNC.det','condition'] = 'UNC det.'
df.loc[df['condition'] == 'AUdet.','condition'] = 'AU det.'
df.loc[df['condition'] == 'MS65 BN','condition'] = 'MS65BN'

Оставляем оценки только по Шелдону.

target_list = ['AU58BN', 'AU55', 'F det.', 'AU58', 'AU det.', 'ХF det.', 'MS66RB',\
       'AU53', 'MS60BN', 'MS62BN', 'MS60', 'AU50', 'UNC det.',\
       'MS61BN', 'XF45', 'MS62', 'MS64', 'MS63', 'MS61', 'MS65', 'AU55BN',\
       'MS64RB', 'MS65RB', 'MS64BN', 'PL63', 'MS61RB', 'MS66', 'MS65RD',\
       'MS66RD', 'MS64RD', 'MS63RB', 'MS63BN', 'XF40', 'VF det.', 'MS67',\
       'MS65BN', 'MS66BN', 'VF35', 'VF30', 'MS69', 'MS68',\
       'VF25', 'VG det.', 'F15', 'MS70', 'PF70', 'VF20',\
       'PL65', 'PF64', 'PF69', 'PL61', 'PF60',\
       'PF68', 'PF67', 'PF65', 'PF62', 'PF det.', 'PL62',\
       'AU53BN', 'PL64', 'PL68', 'PL67', 'PL69', 'PF66', 'XF45BN',\
       'XF40BN', 'VF30BN', 'AU50BN', 'VF35BN', 'VF25BN', 'PF63', 'PF63BN',\
       'PF65BN', 'PF64BN', 'PL65RB', 'PF63RD', 'MS62RB', 'MS63RD',\
       'PL64BN', 'MS61RD', 'PL66', 'PF61', 'VG8', 'F12', 'PF67RD',\
       'F15BN', 'PL65RD', 'MS62RD', 'AU58RB', 'PL70', 'PF64RB',\
       'PL60', 'PF61BN', 'PF66RD', 'G det.', 'PF63RB', 'PF64RD',\
       'MS60RB', 'MS67RD', 'PF62BN', 'АU58', 'VG10', 'АU55', 'VF20BN',\
       'PF62RB', 'MS65RВ', 'PL65BN', 'PF60BN',\
       'PL66RB', 'PF65RD', 'PL58', 'PL63RB',\
       'PF65RB', 'PF66BN', 'PF58', 'G4', 'G6', 'PF55', 'PL64RB',\
       'MS67RB', 'AU55RB', 'F12BN', 'PF53']

df = df.loc[df['condition'].isin(target_list)]
len(df['condition'].unique())

Добавляем имена изображений и столбец о наличии det.

df['det'] = df['condition'].str.find('det')>0
df['revers'] = df['number']+'r.jpg'
df['avers'] = df['number']+'a.jpg'
df.info()

Сохраняем конечный датафрейм с которым будем работать.

df.to_csv(r'D:\Python\coin_data_final.csv')
